* Added --no-compile flag to m2setup.sh and updated help info
* Added php-zip Magento 2 dependency to web role provisioner - resolves #79
* Added install of ack to node provisioner
* Added filters to bootstrap provisioner to quite the output of ssl private key gen
* Updated composer install to dynamically use latest snapshot
* Fixed bug breaking m2setup running within PHP 7 node
* Removed PHP 5.3 support from the environment

+++ vagrant/guest/bin/m2setup.sh
@@ -36,6 +36,7 @@ test -z $ADMIN_PASS && ADMIN_PASS="$(openssl rand -base64 24)"
 # user set flags
 SAMPLEDATA=
 ENTERPRISE=
+NO_COMPILE=
 GITHUB=
 VERBOSE=
 
@@ -54,6 +55,9 @@ for arg in "$@"; do
             ;;
         -g|--github)
             GITHUB=1
+            ;;
+        -C|--no-compile)
+            NO_COMPILE=1
             ;;
         --hostname=*)
             HOSTNAME="${arg#*=}"
@@ -128,15 +132,16 @@ for arg in "$@"; do
             echo "  -d : --sampledata                       triggers installation of sample data"
             echo "  -e : --enterprise                       uses enterprise meta-packages vs community"
             echo "  -g : --github                           will install via github clone instead of from meta-packages"
-            echo "       --proj-version=<proj-version>                if installing via Composer, what version of Magento should be installed"
+            echo "  -C : --no-compile                       skips DI compilation process and static asset generation"
+            echo "       --proj-version=<proj-version>      composer package version to use during installation"
             echo "       --hostname=<hostname>              domain of the site (required input)"
             echo "       --urlpath=<urlpath>                path component of base url and install sub-directyory"
             echo "       --branch=<branch>                  branch to build the site from (defaults to develop)"
             echo "       --backend-frontname=<frontname>    alphanumerical admin username (defaults to backend)"
             echo "       --admin-user=<admin>               alphanumerical admin username (defaults to admin)"
-            echo "       --admin-email=<email>              admin account email address (required input)"
-            echo "       --admin-first=<name>               admin user first name (required input)"
-            echo "       --admin-name=<name>                admin user last name (required input)"
+            echo "       --admin-email=<email>              admin account email address"
+            echo "       --admin-first=<name>               admin user first name"
+            echo "       --admin-name=<name>                admin user last name"
             echo ""
             exit -1
             ;;
@@ -193,9 +198,9 @@ if [[ -f /etc/.vagranthost ]]; then
     exit 1
 fi
 
-php_vercheck=$(php -r 'echo version_compare(PHP_VERSION, "5.5.0", ">=") ? 1 : "";')
+php_vercheck=$(php -r 'echo version_compare(PHP_VERSION, "5.6.0", ">=") ? 1 : "";')
 if [[ -z $php_vercheck ]]; then
-    >&2 echo "Error: Magento 2 requires PHP 5.5 or newer"
+    >&2 echo "Error: Magento 2 requires PHP 5.6 or newer"
     exit
 fi
 
@@ -337,6 +342,9 @@ else
     install_from_packages
 fi
 
+# link session dir so install won't choke trying to lock a session file on an nfs mount
+ln -s /var/lib/php/session var/session
+
 # either install or upgrade database
 print_info_flag=
 code=
@@ -371,17 +379,19 @@ else
     bin/magento setup:upgrade $NOISE_LEVEL
 fi
 
-echo "==> Recompiling DI and static content"
-rm -rf var/di/ var/generation/
-# Magento 2.0.x required usage of multi-tenant compiler (see here for details: http://bit.ly/21eMPtt).
-# Magento 2.1 dropped support for the multi-tenant compiler, so we must use the normal compiler.
-if [ `bin/magento setup:di:compile-multi-tenant --help &> /dev/null; echo $?` -eq 0 ]; then
-    bin/magento setup:di:compile-multi-tenant $NOISE_LEVEL
-else
-    bin/magento setup:di:compile $NOISE_LEVEL
+if [[ ! $NO_COMPILE ]]; then
+    echo "==> Compiling DI and generating static content"
+    rm -rf var/di/ var/generation/
+    # Magento 2.0.x required usage of multi-tenant compiler (see here for details: http://bit.ly/21eMPtt).
+    # Magento 2.1 dropped support for the multi-tenant compiler, so we must use the normal compiler.
+    if [ `bin/magento setup:di:compile-multi-tenant --help &> /dev/null; echo $?` -eq 0 ]; then
+        bin/magento setup:di:compile-multi-tenant $NOISE_LEVEL
+    else
+        bin/magento setup:di:compile $NOISE_LEVEL
+    fi
+    bin/magento setup:static-content:deploy $NOISE_LEVEL
+    bin/magento cache:flush $NOISE_LEVEL
 fi
-bin/magento setup:static-content:deploy $NOISE_LEVEL
-bin/magento cache:flush $NOISE_LEVEL
 
 echo "==> Reindexing and flushing magento cache"
 bin/magento indexer:reindex $NOISE_LEVEL
